mlbench.2dnormals {mlbench}	R Documentation

2-dimensional Gaussian Problem

Description

Each of the cl classes consists of a 2-dimensional Gaussian. The centers are equally spaced on a circle around the origin with radius r.

Usage

mlbench.2dnormals(n, cl=2, r=sqrt(cl), sd=1)

Arguments

n
number of patterns to create

cl
number of classes

r
radius at which the centers of the classes are located

sd
standard deviation of the Gaussians

Value

Returns an object of class "bayes.2dnormals" with components

x
input values

classes
factor vector of length n with target classes

Examples

# 2 classes
p <- mlbench.2dnormals(500,2)
plot(p)
# 6 classes
p <- mlbench.2dnormals(500,6)
plot(p)

[Package mlbench version 2.1-1 Index]