mlbench.cassini {mlbench}	R Documentation

Cassini: A 2 Dimensional Problem

Description

The inputs of the cassini problem are uniformly distributed on a 2-dimensional space within 3 structures. The 2 external structures (classes) are banana-shaped structures and in between them, the middle structure (class) is a circle.

Usage

mlbench.cassini(n, relsize=c(2,2,1))

Arguments

n
number of patterns to create

relsize
relative size of the classes (vector of length 3)

Value

Returns an object of class "mlbench.cassini" with components

x
input values

classes
vector of length n with target classes

Author(s)

Evgenia Dimitriadou and Andreas Weingessel

Examples

p <- mlbench.cassini(5000)
plot(p)

[Package mlbench version 2.1-1 Index]