mlbench.circle {mlbench}	R Documentation

Circle in a Square Problem

Description

The inputs of the circle problem are uniformly distributed on the d-dimensional cube with corners {+-1}. This is a 2-class problem: The first class is a d-dimensional ball in the middle of the cube, the remainder forms the second class. The size of the ball is chosen such that both classes have equal prior probability 0.5.

Usage

mlbench.circle(n, d=2)

Arguments

n
number of patterns to create

d
dimension of the circle problem

Value

Returns an object of class "mlbench.circle" with components

x
input values

classes
factor vector of length n with target classes

Examples

# 2d example
p<-mlbench.circle(300,2)
plot(p)
#
# 3d example
p<-mlbench.circle(300,3)
plot(p)

[Package mlbench version 2.1-1 Index]