mlbench.cuboids {mlbench}	R Documentation

Cuboids: A 3 Dimensional Problem

Description

The inputs of the cuboids problem are uniformly distributed on a 3-dimensional space within 3 cuboids and a small cube in the middle of them.

Usage

mlbench.cuboids(n, relsize=c(2,2,2,1))

Arguments

n
number of patterns to create

relsize
relative size of the classes (vector of length 4)

Value

Returns an object of class "mlbench.cuboids" with components

x
input values

classes
vector of length n with target classes

Author(s)

Evgenia Dimitriadou, and Andreas Weingessel

Examples

p <- mlbench.cuboids(7000)
plot(p)
## Not run: 
library(Rggobi)
g <- ggobi(p$x)
g$setColors(p$class)
g$setMode("2D Tour")

## End(Not run)

[Package mlbench version 2.1-1 Index]