mlbench.hypercube {mlbench}	R Documentation

Corners of Hypercube

Description

The created data are d-dimensional spherical Gaussians with standard deviation sd and means at the corners of a d-dimensional hypercube. The number of classes is 2^d.

Usage

mlbench.hypercube(n=800, d=3, sides=rep(1,d), sd=0.1)
hypercube(d)

Arguments

n
number of patterns to create

d
dimensionality of hypercube, default is 3

sides
lengths of the sides of the hypercube, default is to create a unit hypercube

sd
standard deviation

Value

Returns an object of class "mlbench.hypercube" with components

x
input values

classes
factor of length n with target classes

Examples

p <- mlbench.hypercube()
plot(p)

library("lattice")
cloud(x.3~x.1+x.2, groups=classes, data=as.data.frame(p))

[Package mlbench version 2.1-1 Index]