mlbench.peak {mlbench}	R Documentation

Peak Benchmark Problem

Description

Let r=3u where u is uniform on [0,1]. Take x to be uniformly distributed on the d-dimensional sphere of radius r. Let y=25exp(-.5r^2). This data set is not a classification problem but a regression problem where y is the dependent variable.

Usage

mlbench.peak(n, d=20)

Arguments

n
number of patterns to create

d
dimension of the problem

Value

Returns a list with components

x
input values (independent variables)

y
output values (dependent variable)

[Package mlbench version 2.1-1 Index]