mlbench.threenorm {mlbench}	R Documentation

Threenorm Benchmark Problem

Description

The inputs of the threenorm problem are points from two Gaussian distributions with unit covariance matrix. Class 1 is drawn with equal probability from a unit multivariate normal with mean (a,a,…,a) and from a unit multivariate normal with mean (-a,-a,…,-a). Class 2 is drawn from a multivariate normal with mean at (a,-a,a, …,-a), a=2/d^{0.5}.

Usage

mlbench.threenorm(n, d=20)

Arguments

n
number of patterns to create

d
dimension of the threenorm problem

Value

Returns an object of class "mlbench.threenorm" with components

x
input values

classes
factor vector of length n with target classes

References

Breiman, L. (1996). Bias, variance, and arcing classifiers. Tech. Rep. 460, Statistics Department, University of California, Berkeley, CA, USA.

Examples

p<-mlbench.threenorm(1000, d=2)
plot(p)

[Package mlbench version 2.1-1 Index]