mlbench.waveform {mlbench}	R Documentation

Waveform Database Generator

Description

The generated data set consists of 21 attributes with continuous values and a variable showing the 3 classes (33% for each of 3 classes). Each class is generated from a combination of 2 of 3 "base" waves.

Usage

  mlbench.waveform(n)

Arguments

n
number of patterns to create

Value

Returns an object of class "mlbench.waveform" with components

x
input values

classes
factor vector of length n with target classes

Source

The original C code for the waveform generator hase been taken from the UCI Repository of Machine Learning Databases at

ftp://ftp.ics.uci.edu/pub/machine-learning-databases

http://www.ics.uci.edu/~mlearn/MLRepository.html

The C code has been modified to use R's random number generator by Friedrich Leisch, who also wrote the R interface.

References

Breiman, L. (1996). Bias, variance, and arcing classifiers. Tech. Rep. 460, Statistics Department, University of California, Berkeley, CA, USA.

Newman, D.J. & Hettich, S. & Blake, C.L. & Merz, C.J. (1998). UCI Repository of machine learning databases [http://www.ics.uci.edu/~mlearn/MLRepository.html]. Irvine, CA: University of California, Department of Information and Computer Science.

Examples

  p<-mlbench.waveform(100)
  plot(p)

[Package mlbench version 2.1-1 Index]