plot.mlbench {mlbench}	R Documentation

Plot mlbench objects

Description

Plots the data of an mlbench object using different colors for each class. If the dimension of the input space is larger that 2, a scatter plot matrix is used.

Usage

## S3 method for class 'mlbench'
plot(x, xlab="", ylab="", ...)

Arguments

x
Object of class "mlbench".

xlab
Label for x-axis.

ylab
Label for y-axis.

...
Further plotting options.

Examples

# 6 normal classes
p <- mlbench.2dnormals(500,6)
plot(p)

# 4-dimensiona XOR
p <- mlbench.xor(500,4)
plot(p)

[Package mlbench version 2.1-1 Index]